requirements.txt
fastapi
uvicorn
pandas
numpy
requests
openpyxl
python-multipart
python-dotenv
supabase
google-generativeai
reportlab
